Bring out the common code and integrate it
