+++ src/Server/RestRouterHandlers/ColumnRestRouterHandler.h
@@ -22,11 +22,7 @@ private:
     String executePatch(const Poco::JSON::Object::Ptr & payload, Int32 & http_status) const override;
 
 private:
-    String buildResponse() const;
-    String processQuery(const String & query, Int32 & http_status) const;
     bool columnExist(const String & database_name, const String & table_name, const String & column_name) const;
-
-    ASTPtr parseQuerySyntax(const String & create_table_query) const;
 };
 
 }
