+++ src/Server/ya.make
@@ -33,6 +33,8 @@ SRCS(
     ProtocolServerAdapter.cpp
     ReplicasStatusHandler.cpp
     RestHTTPRequestHandler.cpp
+    RestRouterHandlers/ColumnRestRouterHandler.cpp
+    RestRouterHandlers/CommonUtils.cpp
     RestRouterHandlers/DatabaseRestRouterHandler.cpp
     RestRouterHandlers/IngestRawStoreHandler.cpp
     RestRouterHandlers/IngestRestRouterHandler.cpp
